Add cards, suggestions, lists, and carousels.
Cards and suggestions are added to the simple ask/tell class instance.
However, lists and carousels spawn a new instance inhereting
from the _CardWithItems class.

This was done so that items can easily be added with the add_item
method (which is not exposed for the ask/tell classes).

To create lists and carousels as method callswould require long chains
of method calls.

Perhaps revist this implementation