Revert SSML/XML escaping
Reverting the entire SSML/XML escape mechanism. Saxutils escape is meant to
escape strings that go in XML strings, not XML strings themselves. As
such it now escapes the entire response, including the valid SSML
directives.

This still leaves the problem that SSML responses must be escaped
somewhere, but this has to happen before rendering the template, not
after it, and thus has to be left to the user.

I'm very sorry for the confusion.