Log only the function name instead of its address in memory
The memory address is an implementation detail that almost never helps
during debugging. It does however pose a challenge for tools that
analyze and organize exceptions based on the hash of their stacktrace
(e.g. Sentry). Removing it helps in those situations by not reporting the
same problem over and over again.